Restructure HmsReportWriter. Global Parameter in progress
